Stable App 2

+++ app/src/main/java/com/ocp/evalformation/ui/rh/import_data/ImportViewModel.kt
@@ -678,7 +678,7 @@ class ImportViewModel @Inject constructor(
                         debut                 = debut,
                         fin                   = fin,
                         Formateur             = formateur,
-                        dateAppreciation      = dateAppreciation.toString(),
+                        dateAppreciation      = dateAppreciation.toInt().toString(),
                         entite                = entite,
                         categorie             = categorie,
                         division              = division,
